refactor: move several to module `math`
