+++ qulab/math/__init__.py
@@ -1,0 +1,3 @@
+from qulab.math.func import (get_unit_prefix, skew, kurtosis,
+                get_probility, threshold, get_threshold_visibility,
+                FWHM_of_normal_distribution, Std_of_norm_from_FWHM)
